getting-started-guides/python/requirements.txt
blinker==1.8.2
certifi==2024.7.4
charset-normalizer==3.3.2
click==8.1.7
Deprecated==1.2.14
Flask==3.0.3
googleapis-common-protos==1.63.0
idna==3.7
importlib-metadata==7.0.0
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
protobuf==4.25.3
requests==2.32.3
setuptools==70.0.0
typing_extensions==4.12.0
urllib3==2.2.2
Werkzeug==3.0.3
wrapt==1.16.0
zipp==3.19.1

opentelemetry-exporter-otlp-proto-http==1.26.0
opentelemetry-distro==0.47b0